# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "43.200.170.45",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/signup": {
      "post": {
        "tags": [
          "회원가입 API"
        ],
        "summary": "회원가입 API",
        "description": "-",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userEmail": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "회원 확인용 API"
        ],
        "summary": "회원 확인용 API",
        "description": "-",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "로그인 API"
        ],
        "summary": "로그인 패스워드 API",
        "description": "-",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userEmail": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/users/search": {
      "get": {
        "tags": [
          " 유저 검색 API"
        ],
        "summary": "유저 검색 API",
        "description": "-",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/mailing": {
      "post": {
        "tags": [
          " 인증코드 메일링 API"
        ],
        "summary": "인증코드 메일링 API",
        "description": "-",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userEmail": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}